requirements.txt
Pillow==9.0.1
tk==0.1.0
tkinterdnd2-0.3.0